* clean test suite * Added dynamic checks for method guards (partially solves XTENLANG-2015)
The design is as follows:
Added to ProcedureInstance:
    boolean checkGuardAtRuntime();
    X10ProcedureInstance<T> checkGuardAtRuntime(boolean check);
Implement them in ProcedureInstance_c and in MacroType_c
In Matcher :
                    X10CompilerOptions opts = (X10CompilerOptions) context.typeSystem().extensionInfo().getOptions();
                    if (!opts.x10_config.STATIC_CALLS)
                        newMe = newMe.checkGuardAtRuntime(true);
                    else
                        throw new SemanticException("Call invalid; calling environment does not entail the method guard.");

This will be automatically propagated to X10Call_c (that has a MethodInstance).
No need to change tryImplicitConversions or findMethod because they return Pair<MethodInstance,...>

Found all the places where a procedure or method instance are reported for their .error()  and in those places I also reported if a dynamic check was added to the method guard. (like in X10Call_c, X10ConstructorCall_c, X10Binary_c, etc)

Code generation will need to look at checkGuardAtRuntime() and generate code appropriately (see XTENLANG-2330).
(after flattenning it is easy, but before flattenning it would require a closure that evaluates "this" and the arguments first).