Fairly major overhaul of the build process to enable precompiled X10 libraries to be supported by the C++ backend.
All of the special logic related to the standard lib in the 
compiler has been generalized to work on an arbitrary set of
libraries, specified on the command line with a new -x10lib
option.  The default standard lib is provided to x10c/x10c++
as an x10lib option within the script.  

Build process was changed so that artifacts related to the 
stdlib get installed in x10.dist/stdlib instead of scattered
through x10.dist.