Preparation for constructor call inlining, part 2.
Refactoring/simplification of the call hierarchy: all method, constructor, and closure calls now implement InlinableCall.

(Note: this means that two polyglot interfaces now inherit from an X10 interface.  I believe this is consistent with ongoing efforts to merge polyglot into X10.  If not, we will need to revert this and introduce a more complicated restructuring.)