Teach the constraint solver that if f.a != g.a, then f!=g.
So when adding the disequality between f.a and g.a add it between f an g as well.

Conversely, when looking to establish f !=g, if you can establish f.x == a  (for some x), and g.x ==b and a !=b, then conclude f !=g.

This fixes XTENLANG-2450.