Added a flag named X10_NOWRITEBUFFER.  When this flag is set, the sockets operate as they did before this commit, with no intermediate buffering between places.  This means that the sending side can only send until the TCP buffer is full between the sender and receiver, and when this limit is reached, the sender is blocked until the receiver reads in some data, making space in the network buffers.  This can potentially lead to deadlock on systems with small network buffers, such as seen in XTENLANG-2572.
By default this flag is not set.  This new behavior turns on non-blocking sockets between places, and when the network buffer fills up, the sending side detects this and buffers the remaining data locally, allowing the X10 program to continue.  The sockets transport regularly checks for buffered data and attempts to flush it with each probe or send operation.  This prevents the potential deadlock above, but it's possible for a place to buffer up large amounts of data as it runs ahead of the reading side, using up local memory, potentially crashing the local program when the memory runs out.  Thus, the flag to disable it.