Major simplification of C++ codegen for structs.  Eliminate .struct_h files and dual class (data/methods) strategy and more closely align with the codegen structure for classes.
The previous codegen strategy was based on a misunderstanding 
in which it was thought that in order to pass a struct by value
it was necessary to have a much stronger separation between the
data definition and the method definitions.  This was wrong,
so we can use a much simpler and more natural codegen.