Fix for the recent regression caused by changes in StaticInitializer
Before StaticInitializer changes Console.{OUT,ERR,IN} were detected
as per-place initializable (due to seemingly incorrect assumption
that any call to java-native method is okay for per-place initialization
with obvious counter-example System.nanoTime()).
Recent changes disabled the deep search, but caused OUT,ERR,IN
to be initialized in place-0, and subsequently broadcasted to other places.
Which was exactly the point of failure, since PrintStream cannot
be serialized. This caused massive failures on multi-vm configuration.