XTENLANG-2640: Problem in stream ordering that showed up in the presence of implicit inlining, but was actually latent in the code even without it.
Eliminate a couple places in the codegenerator outside of processClass where
a stream explicitly in the .CC file was being created and instead create 
a specific stream for static data declarations.  This ensures that the streams
will be properly ordered and the include of the class header file will be the
very first non-comment line in the .cc file.