Fix XTENLANG-2711.
I had earlier changed the way information is pushed into the context stack in the body of classes. Now the class invariant and the type bounds for type parameters of a class are not pushed into the context associated with the class declaration. For, this information is not in scope in some members in the class body, such as static members and constructors.

Instead this information is pushed onto the context stack in the bodies of instance fields and instance methods. 

Unfortunately, the code for looking up fields assumed that this information is available in the class context. Hence, this code has been changed. A new context argument, lookupContext is sent in to findVariableSilent and friends. This is typically a  context lower in the context stack than the class context, and contains the class invariant and type bounds information (if the context is associated with the body of an instance methor or an instance field).