Fix XTENLANG 2389.
5/22/2011 8:20:50 AM

Handling outer this correctly.

The basic design is that we introduce the XVar QualifiedVar(Type,
XVar) in the constraint system. 

This represents the outer term Type.XVar.

That is for every value v of an inner class A.B...X.Y, there is a
distinct value U.v for every outer class U in the sequence A, A.B,
A.B.C, ..., A.B..X. The type of U.v is U. 

Note that v may be any XVar, including self or this.

Now during compilation there are two distinct representations of "outer
this" variables.

First, the context maintains a stack of class scopes, corresponding to
each nested class declaration that is entered by a visitor. Each class
scope C has an associated "this" variable. This is typically an
instance of CThis, with a type field containing C. In the innermost
scope, the "implicit representation" of a qualified this, Q.this, is
the "this" variable associated with the class scope for Q.

Second, we introduce QualifiedVar's as described above. The
QualifiedVar term new QualifiedVar(Q, t) is an "explicit
representation" of Q.this (where t represents this, typically t is a
CThis instance). That is, if this is of type F, then Q.this is
represented as a QualifiedVar(Q, t), where t is an instance of CThis
with the associated type F. This will print as "Q.this(:F)". 

The advantage of an explicit representation is that it permits
substitution. The substitution of a term s for t in QualifiedVar(Q,t)
is a new term QualifiedVar(Q, s). This implements the idea that the
subtitution of, say, x for self in A.self is the term A.x. (Such a
term represents the "outer" object at type A associated with x --
assuming that x is of some inner type.)

Note that currently X10 does not support "outer self" or "outer
variable" syntax, i.e. expressions such as A.self or A.x, where x is a
local variable of an inner type. This syntax will be supported in the
future. The programmer now has to use nested types to get access to
"outer self", e.g. write A{self.i==3}.B{self.j=="x"}. Internally such
a type is represented with the constraints flattened and stored at the
innermost level. That is, the type is represented as
A.B{A.self.i==3,self.j=="x"}.

The current code commit maintains a correspondence between the two
representations by setting the type of the X10Special_c term Q.this to
be Q{self==explicit rep, self==implicit rep}. 

It also ensures that when a new object is created constraints from the
outer object are transferred to the type of the object being created.

Added "specialAsQualifiedVar" methods on Context to permit a user of
the context to specify whether this variables shold be resolved using
implicit or explicit representation. The default is implicit
representation.

The bodies of properties are always translated with explicit
representation.

Also improved error messages for invalid parameter errors.

Made Main.loadExtension public so that junit can load an extension (so
that constraints tests involving types can be written).

Also threading the base type of the constraint through the Synthesizer
so that code can be geneerted for self. (This may actually not be
needed.) 

The synthesizer now instantiates self in the constraint for a
parameter with the local variable being used to stand for the
parameter.