Final piece of fix for XTENLANG-2419: Performance of Array.reduce()
The optimized path for reduce/scan/fill needs to test rect, not region.rect.
In principle it shouldn't matter, but it looks like the entailment of 
rect==>region.rect implied by the class invariant orect==region.rect either
isn't being represented in the type of the Array, or the constant propagator
isn't asking the question in a way that would cause it to come into play.

In any case, testing rect is better (one fewer load when isn't eliminated
at compile time).

With this change, I see identical performance for the two loop nests.

[dgrove@wahtutca myTests]$ ./a.out 100000000
vector size 100000000
loop max is 1: 388.929 ms
reduce max is 1: 383.795 ms