Improved CodeCleanUp pass to eliminate some unnecessary breaks
The added transformation transforms the labeled blocks
of the form

    label: {
	stmts;
	break label;
    }

to the simpler block

    {
	stmts;
    }

provided that `break label` was the only use of the label.

The labeled break are produced by Inliner by mechanically
converting `return` to `break`, so this pattern
may occur quite often. It is translated to GOTO
bytecodes in class file, which in turn may affect
the heuristics JIT uses to choose optimization level
for the method.