Don't try to generate a #include for a Java type.
If the program actually needed the #include (ie, really used the type)
then it will still fail at post-compilation time.
However, if the type was only actually used in a type constraint, 
then it can still be compiled successfully by the postcompiler 
if we just ignore the use of the Java type.