XTENLANG-2959 : Bug in deserialization of Aliased org.hadoop.Writable.
Conservative fix of calling jds.update_reference.  If possible, we should
instead avoid creating an extra instance of the object by not calling
newInstance in the deserialization routine.  We've already created an 
instance of the object via Unsafe in the caller.  So, the way the code is now
is sub-optimal unless the semantics of Hadoop serialization require the
constructor to be called.