Import 2.2.1 version of LU from benchmarks/hpcc and convert to X10DT C++ backend project.
Note to actually execute the program BLAS must be installed on the target machine.  The default project setup assumes adding -lblas to the g++ linking command will suffice to find the blas libraries.  Adjustments to x10_platform.conf may be needed if this is not true on a particular machine.