Add support for pre-assigning serialization ids for commonly serialized types.  Currently support a built-in list of types and a property that the user may set on the command line to add more.
One key limitation is that there isn't yet a robust way to specify
key closure types that are used in the implementation of at/finish/etc
in XRX.  (No stable mapping of the X10 closure to the name for the
generated class).  Will need to optimize this case sometime fairly soon.