WIP on serialization improvements for M3R.
Switch Managed X10 serialization/deserialization wire format to
use an "inline" meta-data dictionary.  This enables us to write
both the data and the meta-data into a single stream, which in
turns opens up the ability to have the Serializer wrapping 
a FileOutputStream or other more direct I/O channel to allow
the serialized bytes to stream directly to their destination.