Replaced the static fields Place.MAX_PLACES and Place.ALL_PLACES with method calls, so that any call will be threaded through to the central Hazelcast reference for the number of places.
This includes changes in code that makes use of MAX_PLACES, which is frequently used.  I converted all of these references to method calls.  THIS IS NOT ALWAYS THE CORRECT THING TO DO, as much of these uses will not work correctly if the value changes.  Much of this needs to be examined more closely for Elastic X10, but for now, this gets it to compile.