XTENLANG-3396: Native X10 should not emit using declarations for static methods.
When a subclass defines a method with the same name as a superclass 
static method, we should not emit a using declaration to bring the 
superclass's static method into the namespace of the subclass.  
We invoke static methods via FQN so it is not needed in the subclass
namespace.  Furthermore in the case of generic classes, the static method 
is generated into the void specialization of the class and therefore is not 
even present at the C++ level in the primary superclass (thus causing the 
using declaration in the subclass to be reported as a C++ compilation error).