XTENLANG-3463: Should not need to force BUSY_WAITING in Resilient X10
Change @Immediate worker thread to also check for a chance in the number
of known dead Places and schedule an activity to notifyPlaceDeath when 
the number of dead Places changes.  When busy waiting is disabled, 
all of the non-Immediate workers may be parked as spare threads (if the
Place is fully idle).  Therefore we need to trigger processing of Place
death events from the @Immediate worker.