WIP. Managed mode works now with single precision. Checked with FwdBwd.
Next up: Scripts need to be updated to make it easy to build single or double precision versions of (managed, native) GML. Applications need to be moved to use ElemType so they can work in both single and double precision made (the only thing that changes is which GML library you link with). Tests need to be run. Comments need to be updated.