New parser (manual merge of the parser branch).
The new parser introduces non backward compatible syntax changes:
- there is no implicit return in method definitions;
- the body of method and constructor definitions cannot be a block introduced with an "=".

Therefore, the following method definition is now incorrect:
 def f (x: long) = {
     x + 1
 }
It should be rewritten:
 def f (x: long) {
    return x + 1;
 }
or:
 def f (x: long) = x + 1;

Other change: any expression can be used as a statement.

To use the parser you have to set the option -ANTLR_PARSER.
The option -DISPLAY_PARSE_TREE shows in a window the parsed program.