Bugfixes in the validation of put/get operations.  regular (non-put/get) messages transmit data in the x10rt_msg_params structure, but put/get tests set that structure to null, and use an additional set of buffer and len arguments for transmitting the data.  The validation inside of recv_get/put_ping/pong always saw a p->len of 0, and so never validated the buffer contents.
There is still a problem in the test case when running on RDMA hardware.  When validation is enabled, the destination buffer is reset with zeros after validating, so that the next put/get can be verified to replace those with the correct contents (0,1,2,3...).  On sockets this works fine because the single-threaded operation of the test ensures that the reset occurs before the next put/get is processed and real data is filled in.  But under RDMA, we loose the single-threaded nature of the test program because the remote place may write to the destination before the buffer is reset, causing validation after that to fail.  The only way to prevent this is to use the "window" argument to the program, setting the window size to 1.